TicTacToe - Human Vs Cpu
Implemented an algorithm that weighs the winning patterns the human plays on the board, sorts them by weight, and then checks if a spot in the pattern is free, in which it places a 'O'. If for some reason all spaces on that pattern are taken, it will move to the next highest weighted pattern and check again. 'It' doesnt aim to win, just block any wins.